Separate native configuration runtime application
