+++ components/espcontrol/panel_config_esphome_text.h
@@ -5,11 +5,11 @@
 
 #include "esphome/components/text/text.h"
 
-#include "panel_config_legacy_adapter.h"
+#include "panel_config_text_bindings.h"
 
 namespace espcontrol::configuration {
 
-class EspHomeLegacyTextValue final : public LegacyTextValue {
+class EspHomePanelConfigTextValue final : public PanelConfigTextValue {
  public:
   void bind(esphome::text::Text *text) { text_ = text; }
 
@@ -33,5 +33,9 @@ class EspHomeLegacyTextValue final : public LegacyTextValue {
   esphome::text::Text *text_{nullptr};
   std::string empty_;
 };
+
+// Temporary source-compatible name while device wiring moves away from the
+// legacy configuration path.
+using EspHomeLegacyTextValue = EspHomePanelConfigTextValue;
 
 }  // namespace espcontrol::configuration
