+++ dev-docs/check-matrix.md
@@ -26,9 +26,9 @@ This detailed routing table is generated directly from `scripts/check_tasks_data
 
 | Task | Domains | Parallel-safe | Cache | Cache environment | Cache tools | Declared inputs | Cache-only inputs | Focused command |
 | --- | --- | --- | --- | --- | --- | --- | --- | --- |
-| `firmware-tests` | firmware | Yes | never | — | — | `tests/firmware/**`<br>`components/espcontrol/configuration_service.*`<br>`components/espcontrol/panel_config_service_validator.h`<br>`components/espcontrol/panel_config_capabilities.h`<br>`components/espcontrol/panel_config_capabilities_endpoint.h`<br>`components/espcontrol/panel_config_legacy_adapter.*`<br>`components/espcontrol/panel_config_storage_backend.h`<br>`components/espcontrol/panel_config_espidf_storage.*`<br>`components/espcontrol/espcontrol_app.*`<br>`components/espcontrol/configuration_store.*`<br>`components/espcontrol/panel_config_document.h`<br>`components/espcontrol/button_grid_limits.h`<br>`components/espcontrol/button_grid_slider_geometry.h`<br>`components/espcontrol/button_grid_string.h`<br>`components/espcontrol/button_grid_config_parser.h`<br>`components/espcontrol/button_grid_subpages.h`<br>`components/artwork_image/image_pipeline_policy.h`<br>`common/config/card_normalization_fixtures.json`<br>`common/config/*_card_normalization_fixtures.json`<br>`scripts/generate_saved_config_parser_test.py`<br>`src/webserver/application/config_codec.ts` | — | `python3 scripts/check_tasks.py run-task firmware-tests` |
+| `firmware-tests` | firmware | Yes | never | — | — | `tests/firmware/**`<br>`components/espcontrol/configuration_service.*`<br>`components/espcontrol/panel_config_service_validator.h`<br>`components/espcontrol/panel_config_capabilities.h`<br>`components/espcontrol/panel_config_capabilities_endpoint.h`<br>`components/espcontrol/panel_config_legacy_adapter.*`<br>`components/espcontrol/panel_config_runtime_adapter.*`<br>`components/espcontrol/panel_config_text_bindings.*`<br>`components/espcontrol/panel_config_storage_backend.h`<br>`components/espcontrol/panel_config_espidf_storage.*`<br>`components/espcontrol/espcontrol_app.*`<br>`components/espcontrol/configuration_store.*`<br>`components/espcontrol/panel_config_document.h`<br>`components/espcontrol/button_grid_limits.h`<br>`components/espcontrol/button_grid_slider_geometry.h`<br>`components/espcontrol/button_grid_string.h`<br>`components/espcontrol/button_grid_config_parser.h`<br>`components/espcontrol/button_grid_subpages.h`<br>`components/artwork_image/image_pipeline_policy.h`<br>`common/config/card_normalization_fixtures.json`<br>`common/config/*_card_normalization_fixtures.json`<br>`scripts/generate_saved_config_parser_test.py`<br>`src/webserver/application/config_codec.ts` | — | `python3 scripts/check_tasks.py run-task firmware-tests` |
 | `web-unit` | web | Yes | deterministic | — | `node` | `tests/web/unit/**`<br>`tests/web/*.test.ts`<br>`src/webserver/**`<br>`product/v2/product_compatibility.json`<br>`compatibility/fixtures/panel_config_migration_v1.json`<br>`devices/manifest.json`<br>`scripts/load_typescript_module.js`<br>`package-lock.json` | — | `python3 scripts/check_tasks.py run-task web-unit` |
-| `mutations` | firmware, web | No | never | — | — | `tests/mutations/**`<br>`tests/firmware/**`<br>`tests/web/*.test.ts`<br>`tests/web/unit/**`<br>`components/espcontrol/configuration_service.*`<br>`components/espcontrol/panel_config_service_validator.h`<br>`components/espcontrol/panel_config_legacy_adapter.*`<br>`components/espcontrol/panel_config_storage_backend.h`<br>`components/espcontrol/panel_config_espidf_storage.*`<br>`components/espcontrol/configuration_store.*`<br>`components/espcontrol/panel_config_document.h`<br>`src/webserver/model/**`<br>`src/webserver/generated/**`<br>`scripts/run_mutations.py` | — | `python3 scripts/check_tasks.py run-task mutations` |
+| `mutations` | firmware, web | No | never | — | — | `tests/mutations/**`<br>`tests/firmware/**`<br>`tests/web/*.test.ts`<br>`tests/web/unit/**`<br>`components/espcontrol/configuration_service.*`<br>`components/espcontrol/panel_config_service_validator.h`<br>`components/espcontrol/panel_config_legacy_adapter.*`<br>`components/espcontrol/panel_config_runtime_adapter.*`<br>`components/espcontrol/panel_config_text_bindings.*`<br>`components/espcontrol/panel_config_storage_backend.h`<br>`components/espcontrol/panel_config_espidf_storage.*`<br>`components/espcontrol/configuration_store.*`<br>`components/espcontrol/panel_config_document.h`<br>`src/webserver/model/**`<br>`src/webserver/generated/**`<br>`scripts/run_mutations.py` | — | `python3 scripts/check_tasks.py run-task mutations` |
 | `generated` | product, firmware, web, docs | Yes | never | — | — | `common/**`<br>`devices/**`<br>`builds/**`<br>`components/espcontrol/**`<br>`src/webserver/**`<br>`compatibility/**`<br>`scripts/build.py`<br>`scripts/build_web_bundle.js`<br>`scripts/web_source.js` | — | `npm run check:generated` |
 | `device-manifest` | product, firmware | Yes | deterministic | — | — | `common/assets/**`<br>`product/v2/device_catalog.json`<br>`devices/**`<br>`builds/**`<br>`scripts/check_device_manifest.py` | — | `python3 scripts/check_tasks.py run-task device-manifest` |
 | `device-manifest-output` | product, firmware, docs | Yes | deterministic | — | — | `common/assets/**`<br>`product/v2/device_catalog.json`<br>`scripts/generate_device_manifest.py` | — | `python3 scripts/check_tasks.py run-task device-manifest-output` |
