+++ scripts/check_tasks_data.py
@@ -88,6 +88,8 @@ TASKS = (
              "components/espcontrol/panel_config_capabilities.h",
              "components/espcontrol/panel_config_capabilities_endpoint.h",
              "components/espcontrol/panel_config_legacy_adapter.*",
+             "components/espcontrol/panel_config_runtime_adapter.*",
+             "components/espcontrol/panel_config_text_bindings.*",
              "components/espcontrol/panel_config_storage_backend.h",
              "components/espcontrol/panel_config_espidf_storage.*",
              "components/espcontrol/espcontrol_app.*",
@@ -110,7 +112,7 @@ TASKS = (
          cache_tools=("node",)),
     task("mutations", ("python3", "scripts/run_mutations.py"),
          domains=("firmware", "web"),
-         inputs=("tests/mutations/**", "tests/firmware/**", "tests/web/*.test.ts", "tests/web/unit/**", "components/espcontrol/configuration_service.*", "components/espcontrol/panel_config_service_validator.h", "components/espcontrol/panel_config_legacy_adapter.*", "components/espcontrol/panel_config_storage_backend.h", "components/espcontrol/panel_config_espidf_storage.*", "components/espcontrol/configuration_store.*", "components/espcontrol/panel_config_document.h", "src/webserver/model/**", "src/webserver/generated/**", "scripts/run_mutations.py"),
+         inputs=("tests/mutations/**", "tests/firmware/**", "tests/web/*.test.ts", "tests/web/unit/**", "components/espcontrol/configuration_service.*", "components/espcontrol/panel_config_service_validator.h", "components/espcontrol/panel_config_legacy_adapter.*", "components/espcontrol/panel_config_runtime_adapter.*", "components/espcontrol/panel_config_text_bindings.*", "components/espcontrol/panel_config_storage_backend.h", "components/espcontrol/panel_config_espidf_storage.*", "components/espcontrol/configuration_store.*", "components/espcontrol/panel_config_document.h", "src/webserver/model/**", "src/webserver/generated/**", "scripts/run_mutations.py"),
          cache="never"),
     task("generated", ("python3", "scripts/build.py", "--check"),
          ("python3", "scripts/build.py", "--self-test"), profiles=PRODUCT,
